chore(web): move `event` utils inside `event/`
