+++ packages/web/src/common/utils/draft/someday.draft.util.test.ts
@@ -2,7 +2,7 @@ import dayjs from "dayjs";
 import { Categories_Event } from "@core/types/event.types";
 import { draftSlice } from "@web/ducks/events/slices/draft.slice";
 import { Activity_DraftEvent } from "@web/ducks/events/slices/draft.slice.types";
-import { assembleDefaultEvent } from "../event.util";
+import { assembleDefaultEvent } from "../event/event.util";
 import { createSomedayDraft } from "./someday.draft.util";
 
 // Mock assembleDefaultEvent since it makes external calls
