+++ packages/web/src/common/utils/draft/someday.draft.util.ts
@@ -4,7 +4,7 @@ import { YEAR_MONTH_DAY_FORMAT } from "@core/constants/date.constants";
 import { Categories_Event } from "@core/types/event.types";
 import { draftSlice } from "@web/ducks/events/slices/draft.slice";
 import { Activity_DraftEvent } from "@web/ducks/events/slices/draft.slice.types";
-import { assembleDefaultEvent } from "../event.util";
+import { assembleDefaultEvent } from "../event/event.util";
 
 /** @deprecated
  * `createSomedayDraft` should not be called outside `useSidebarActions.createSomedayDraft`
