+++ packages/web/src/common/utils/overlap/overlap.test.ts
@@ -1,7 +1,7 @@
 import { Categories_Event } from "@core/types/event.types";
 import { getUserId } from "@web/auth/auth.util";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
-import { assembleDefaultEvent } from "../event.util";
+import { assembleDefaultEvent } from "../event/event.util";
 import { adjustOverlappingEvents } from "./overlap";
 
 jest.mock("@web/auth/auth.util", () => ({
