+++ packages/web/src/ducks/events/sagas/someday.sagas.ts
@@ -2,7 +2,7 @@ import { normalize } from "normalizr";
 import { call, put } from "redux-saga/effects";
 import { Schema_Event } from "@core/types/event.types";
 import { Schema_OptimisticEvent } from "@web/common/types/web.event.types";
-import { handleError } from "@web/common/utils/event.util";
+import { handleError } from "@web/common/utils/event/event.util";
 import { setSomedayEventsOrder } from "@web/common/utils/someday.util";
 import { EventApi } from "@web/ducks/events/event.api";
 import {
