+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/AllDayEvent.tsx
@@ -6,7 +6,7 @@ import {
   ZIndex,
 } from "@web/common/constants/web.constants";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
-import { isOptimisticEvent } from "@web/common/utils/event.util";
+import { isOptimisticEvent } from "@web/common/utils/event/event.util";
 import { getEventPosition } from "@web/common/utils/position/position.util";
 import { Flex } from "@web/components/Flex";
 import { AlignItems, FlexDirections } from "@web/components/Flex/styled";
