+++ packages/web/src/views/Forms/EventForm/EventForm.tsx
@@ -18,7 +18,7 @@ import {
 } from "@web/common/styles/theme.util";
 import { SelectOption } from "@web/common/types/component.types";
 import { mapToBackend } from "@web/common/utils/datetime/web.date.util";
-import { getCategory } from "@web/common/utils/event.util";
+import { getCategory } from "@web/common/utils/event/event.util";
 import { DateControlsSection } from "@web/views/Forms/EventForm/DateControlsSection";
 import { getFormDates } from "@web/views/Forms/EventForm/DateControlsSection/DateTimeSection/form.datetime.util";
 import { RecurrenceSection } from "@web/views/Forms/EventForm/DateControlsSection/RecurrenceSection/RecurrenceSection";
